Add helper class with numbers validation method and simple division operation

+++ MathOperation.cs
@@ -2,19 +2,19 @@ namespace SimpleCalculator
 {
     public class MathOperation {
 
-        public double Addition(int x, int y) {
+        public double Addition(double x, double y) {
             return x + y;
         }
 
-        public double Subtraction(int x, int y) {
+        public double Subtraction(double x, double y) {
             return x - y;
         }
 
-        public double Multiplication(int x, int y) {
+        public double Multiplication(double x, double y) {
             return x * y;
         }
 
-        public double Division(int x, int y) {
+        public double Division(double x, double y) {
             return x / y;
         }
     }
